Localize error enums to modules, better DB error reportng.
